latest_requirements.txt
absl-py==0.8.1
amqp==2.5.1
argcomplete==1.10.0
astor==0.8.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cloudinary==1.18.2
colorama==0.4.1
contractions==0.0.21
cryptography==2.8
Django==2.2.6
django-bootstrap4==1.0.1
django-celery-beat==1.5.0
django-mysql==3.2.0
django-timezone-field==3.1
django-widget-tweaks==1.4.5
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
idna==2.8
IMAPClient==2.1.0
importlib-metadata==0.23
joblib==0.14.0
Keras==2.0.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kombu==4.6.5
log3==0.1.7
lxml==4.4.1
Markdown==3.1.1
mock==3.0.5
more-itertools==7.2.0
mysqlclient==1.4.4
nltk==3.4.5
numpy==1.17.3
olefile==0.46
pandas==0.25.2
pdfminer.six==20181108
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
pyahocorasick==1.4.0
pycparser==2.19
pycryptodome==3.9.0
PyJWT==1.7.1
PyPDF2==1.26.0
python-crontab==2.4.0
python-dateutil==2.8.0
python-pptx==0.6.18
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
schedule==0.6.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
slug==2.0
slugify==0.0.1
sortedcontainers==2.1.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.0.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
textract==1.6.3
textsearch==0.0.17
Theano==1.0.4
tzlocal==1.5.1
unicode-slugify==0.1.3
Unidecode==1.1.1
unsplash-search==0.2.0
urllib3==1.25.6
vine==1.3.0
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.2
zipp==0.6.0
